% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{BreastCancer}
\alias{BreastCancer}
\title{A dataset containing breast cancer data. UCI machine learning repository dataset.}
\format{Entries of patients
\describe{
  \item{class}{benign vs malignant}
  ...
}}
\description{
A dataset containing breast cancer data. UCI machine learning repository dataset.
}
