% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPrivML.R
\docType{package}
\name{diffPrivML}
\alias{diffPrivML}
\alias{diffPrivML-package}
\title{The \code{diffPrivML} package is a collection of machine learning workflow
enhanced with differential privacy.}
\description{
The \code{diffPrivML} package is a collection of machine learning workflow
enhanced with differential privacy.
}
