% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPLogisticRegressionClassifier.R
\name{DPLogisticRegressionClassifier}
\alias{DPLogisticRegressionClassifier}
\title{Implementation of DPLogisticRegressionClassifier. An S3 class.}
\usage{
DPLogisticRegressionClassifier(y, x, lambda, alpha = 0, epsilon = 0)
}
\arguments{
\item{y}{vector with binary outcome}

\item{x}{a data frame with predictor columns}

\item{lambda}{regularization variable}

\item{alpha}{0 for ridge and 1 for lasso}

\item{epsilon}{privacy budget in epsilon-differential private procedure}
}
\value{
a classifier object
}
\description{
Implementation of DPLogisticRegressionClassifier. An S3 class.
}
\examples{
data(iris)
x <- as.matrix(iris[, 1:4])
y <- factor(iris$Species == "setosa")
lrClassifier <- DPLogisticRegressionClassifier(y, x, lambda = 1, alpha = 0, epsilon = 1)

}
\author{
Eva Gong
}
