% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPNaiveBayesClassifier.R
\name{DPNaiveBayesClassifier}
\alias{DPNaiveBayesClassifier}
\title{Implementation of DPNaiveBayesClassifier}
\usage{
DPNaiveBayesClassifier(y, x, mechanism, epsilon)
}
\arguments{
\item{y}{a named vector of outcome labels}

\item{x}{a data frame of named predictors}

\item{mechanism}{function for DF}

\item{epsilon}{privacy budget}
}
\value{
a classifier
}
\description{
Implementation of DPNaiveBayesClassifier
}
\examples{


}
