% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPRandomDecisionTreeClassifier.R
\name{DPRandomDecisionTreeClassifier}
\alias{DPRandomDecisionTreeClassifier}
\title{Implementation of DPRandomDecisionTreeClassifier (S3 class)}
\usage{
DPRandomDecisionTreeClassifier(Y, X, epsilon = NULL, numTrees = 5,
  height = floor(ncol(X)/2))
}
\arguments{
\item{Y}{a named vector of outcome labels}

\item{X}{a data frame with named predictor columns}

\item{epsilon}{privacy budget in epsilon-differential private procedure.
default is non-private.}

\item{numTrees}{# of trees to grow in the ensemble, default is 5}

\item{height}{the maxinum height of each decision tree, default is # of
predictors in X, divided by 2. i.e. floor(ncol(X)/2)}
}
\description{
This is a model that
creates an ensemble of Random Decision Trees for classification.
}
\author{
Terence Tam
}
