% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseFunctions.R
\name{laplace}
\alias{laplace}
\title{Laplace mechanism}
\usage{
laplace(originalData, sensitivity, epsilon)
}
\arguments{
\item{originalData}{data that needs to sensitized}

\item{sensitivity}{sensitivity of the input, captures how much one data point
data can affect output}

\item{epsilon}{privacy budget parameter in epsilon differential private procedure}
}
\value{
the perturbed data
}
\description{
Laplace mechanism
}
