% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseFunctions.R
\name{outputNoise}
\alias{outputNoise}
\title{Output pertubation used by DP logistic regression classifier}
\usage{
outputNoise(n, d, lambda, epsilon)
}
\arguments{
\item{n}{number of rows in the input data}

\item{d}{number of columns in the input data plus constant column}

\item{lambda}{regularization parameter}

\item{epsilon}{privacy budget in epsilon-differential private procedure}
}
\value{
output noise to be added to the coefficients estimates
}
\description{
Output pertubation used by DP logistic regression classifier
}
