% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPLogisticRegressionClassifier.R
\name{predict.DPLogisticRegressionClassifier}
\alias{predict.DPLogisticRegressionClassifier}
\title{predict of logistic regression}
\usage{
\method{predict}{DPLogisticRegressionClassifier}(object, predictX = NULL,
  type = c("classes", "probabilities"), ...)
}
\description{
predict of logistic regression
}
\examples{
data(iris)
x = as.matrix(iris[,1:4])
y = ifelse( iris$Species=="setosa", 1,0 )
lrClassifier = DPLogisticRegressionClassifier(y,x,lambda = 1, alpha = 0, epsilon = 1)
summary(lrClassifier)
result = predict(lrClassifier,type = "classes")
table(result,y)
}
