% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPNaiveBayesClassifier.R
\name{predict.DPNaiveBayesClassifier}
\alias{predict.DPNaiveBayesClassifier}
\title{S3 method for predict}
\usage{
\method{predict}{DPNaiveBayesClassifier}(object, testSetX = NULL, ...)
}
\arguments{
\item{object}{the classifier object for this predict method}

\item{testSetX}{a data frame with named predictor columns}

\item{...}{not applicable for this class}
}
\description{
S3 method for predict
}
\examples{
library(caret)
data(iris)
y <- iris[, 5]
x <- iris[, 1:4]
trainingIndices <-createDataPartition(y, p = 0.70, list = FALSE)
trainingSetX <- x[trainingIndices, ]
trainingSetY <- y[trainingIndices]
testSetX <- x[-trainingIndices, ]
testSetY <- y[-trainingIndices]
naiveBayesDP <- DPNaiveBayesClassifier(trainingSetY, trainingSetX)
predictions <- predict(naiveBayesDP, testSetX)
accuracy <- sum(predictions == testSetY) / length(testSetY)

}
