% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPRandomDecisionTreeClassifier.R
\name{predict.DPRandomDecisionTreeClassifier}
\alias{predict.DPRandomDecisionTreeClassifier}
\title{S3 method for predict}
\usage{
\method{predict}{DPRandomDecisionTreeClassifier}(object, testSetX = NULL, ...)
}
\arguments{
\item{object}{the classifier object for this predict method}

\item{testSetX}{a data frame with named predictor columns}

\item{...}{not applicable for this class}
}
\description{
S3 method for predict
}
\author{
Terence Tam
}
