% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPRandomDecisionTreeClassifier.R
\docType{data}
\name{RandomDecisionTree}
\alias{RandomDecisionTree}
\title{Get random tree}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RandomDecisionTree
}
\description{
Get random tree
}
\keyword{datasets}
