% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPNaiveBayesClassifier.R
\name{summary.DPNaiveBayesClassifier}
\alias{summary.DPNaiveBayesClassifier}
\title{S3 method for DPNaiveBayesClassifier's summary.}
\usage{
\method{summary}{DPNaiveBayesClassifier}(object, ...)
}
\arguments{
\item{object}{the classifier object for this predict method}

\item{...}{not applicable for this class}
}
\description{
S3 method for DPNaiveBayesClassifier's summary.
}
\examples{
data(iris)
y <- iris[, 5]
x <- iris[, 1:4]
naiveBayesDP <- DPNaiveBayesClassifier(y, x)
summary(naiveBayesDP)

}
\author{
Terence Tam
}
