% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPRandomDecisionTreeClassifier.R
\name{summary.DPRandomDecisionTreeClassifier}
\alias{summary.DPRandomDecisionTreeClassifier}
\title{S3 method for summary.}
\usage{
\method{summary}{DPRandomDecisionTreeClassifier}(object, ...)
}
\arguments{
\item{object}{the classifier object for this predict method}

\item{...}{not applicable for this class}
}
\description{
S3 method for summary.
}
\author{
Terence Tam
}
