% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPNaiveBayesClassifier.R
\name{DPNaiveBayesClassifier}
\alias{DPNaiveBayesClassifier}
\title{Implementation of DPNaiveBayesClassifier}
\usage{
DPNaiveBayesClassifier(data, mechanism, budget)
}
\arguments{
\item{data}{a data frame}

\item{mechanism}{function for DF}

\item{budget}{privacy budget}
}
\value{
a trained result
}
\description{
Implementation of DPNaiveBayesClassifier
}
\examples{
1+1

}
