% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPRandomDecisionTreeClassifier.R
\name{DPRandomDecisionTreeClassifier}
\alias{DPRandomDecisionTreeClassifier}
\title{Implementation of DPRandomDecisionTreeClassifier. An S3 class.}
\usage{
DPRandomDecisionTreeClassifier(Y, X, epsilon = NULL, mechanism = NULL,
  numTrees = 5, height = ncol(X)/2)
}
\description{
Implementation of DPRandomDecisionTreeClassifier. An S3 class.
}
