% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPLogisticRegressionClassifier.R
\name{predict.DPLogisticRegressionClassifier}
\alias{predict.DPLogisticRegressionClassifier}
\title{predict method for DP logistic regression}
\usage{
\method{predict}{DPLogisticRegressionClassifier}(object, predictX = NULL,
  type = c("classes", "probabilities"), ...)
}
\arguments{
\item{object}{the classifier object for this predict method}

\item{predictX}{a data frame that holds test data points that will be
predicted. If omitted, data used for model fitting will be used}

\item{type}{the type of prediction. Can be either classes or probabilities}

\item{...}{not applicable for this class}
}
\description{
predict method for DP logistic regression
}
\examples{
data(iris)
x <- as.matrix(iris[, 1:4])
y <- factor(iris$Species == "setosa")
lrClassifier <- DPLogisticRegressionClassifier(y, x, lambda = 1, alpha = 0, epsilon = 1)
summary(lrClassifier)
result <- predict(lrClassifier, type = "classes")
table(result, y)
}
\author{
Yifan Gong
}
