% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPRandomDecisionTreeClassifier.R
\docType{data}
\name{RandomDecisionTree}
\alias{RandomDecisionTree}
\title{This is a factory generator for a RandomDecisionTree R6 class object.
A RandomDeciscionTree, RDT, is a type of decision tree where the partitioning
at each branch is done randomly, unlike other criterion based CART trees.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RandomDecisionTree
}
\description{
To grow a tree, a RandomDecisionTree object can have children pointers to
other RandomDecisionTree objects.
}
\author{
Terence Tam
}
\keyword{datasets}
