% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPLogisticRegressionClassifier.R
\name{summary.DPLogisticRegressionClassifier}
\alias{summary.DPLogisticRegressionClassifier}
\title{summary of logistic regression}
\usage{
\method{summary}{DPLogisticRegressionClassifier}(object, ...)
}
\arguments{
\item{object}{the classifier object for this predict method}

\item{...}{not applicable for this class}
}
\description{
summary of logistic regression
}
\examples{
data(iris)
x <- as.matrix(iris[, 1:4])
y <- factor(iris$Species == "setosa")
lrClassifier <- DPLogisticRegressionClassifier(y, x, lambda = 1, alpha = 0, epsilon = 1)
summary(lrClassifier)

}
\author{
Yifan Gong
}
